Revert "refactor: use `u64::MAX` instead"
This reverts commit 4c53c2cac2d047b59c3880552c9d68ff00108aa6.

+++ contracts/satoshi-bridge/src/api/view.rs
@@ -37,10 +37,14 @@ impl Contract {
     pub fn get_metadata(&self) -> Metadata {
         let root_state = self.data();
         Metadata {
-            super_admins: self.acl_get_super_admins(0, u64::MAX),
-            daos: self.acl_get_grantees(Role::DAO.into(), 0, u64::MAX),
-            operators: self.acl_get_grantees(Role::Operator.into(), 0, u64::MAX),
-            pause_managers: self.acl_get_grantees(Role::PauseManager.into(), 0, u64::MAX),
+            super_admins: self.acl_get_super_admins(0, u64::from(u32::MAX)),
+            daos: self.acl_get_grantees(Role::DAO.into(), 0, u64::from(u32::MAX)),
+            operators: self.acl_get_grantees(Role::Operator.into(), 0, u64::from(u32::MAX)),
+            pause_managers: self.acl_get_grantees(
+                Role::PauseManager.into(),
+                0,
+                u64::from(u32::MAX),
+            ),
             pa_all_paused: self.pa_all_paused(),
             relayer_white_list: root_state.relayer_white_list.iter().cloned().collect(),
             extra_msg_relayer_white_list: root_state
